refactor: merged get by dorm id
